Vision stuff
I made some code. I highly doubt it works, so I commented out everything I made. The code in TargetVisionTape is supposed to get the mean of the Xpos of the contours. Then, if it is not within a certain range, it will make the robot turn until it it within that range. Then the robot should drive straight in order to get to the target. (Also got rid of some deprecated code in the GripPipeline).